+++ src/Entity/Traits/Behavior/Blameable/BlameablePropertiesTrait.php
@@ -1,0 +1,32 @@
+<?php
+
+/**
+ * This file is part of the EightMarq Symfony bundles.
+ *
+ * (c) Norbert Schvoy <norbert.schvoy@eightmarq.com>
+ *
+ * For the full copyright and license information, please view the LICENSE
+ * file that was distributed with this source code.
+ */
+
+declare(strict_types=1);
+
+namespace EightMarq\CoreBundle\Entity\Traits\Behavior\Blameable;
+
+use Doctrine\ORM\Mapping as ORM;
+use Symfony\Component\Security\Core\User\UserInterface;
+
+trait BlameablePropertiesTrait
+{
+    #[ORM\ManyToOne(targetEntity: UserInterface::class)]
+    #[ORM\JoinColumn(onDelete: 'SET NULL')]
+    protected UserInterface|null $createdBy;
+
+    #[ORM\ManyToOne(targetEntity: UserInterface::class)]
+    #[ORM\JoinColumn(onDelete: 'SET NULL')]
+    protected UserInterface|null $updatedBy;
+
+    #[ORM\ManyToOne(targetEntity: UserInterface::class)]
+    #[ORM\JoinColumn(onDelete: 'SET NULL')]
+    protected UserInterface|null $deletedBy;
+}