+++ src/Entity/UuidBasedEntity.php
@@ -14,14 +14,14 @@ declare(strict_types=1);
 namespace EightMarq\CoreBundle\Entity;
 
 use Doctrine\ORM\Mapping as ORM;
+use EightMarq\CoreBundle\Entity\Interfaces\Behavior\BlameableInterface;
 use EightMarq\CoreBundle\Entity\Interfaces\Behavior\SoftDeleteableInterface;
 use EightMarq\CoreBundle\Entity\Interfaces\Behavior\TimestampableInterface;
 use EightMarq\CoreBundle\Entity\Interfaces\UuidBasedEntityInterface;
+use EightMarq\CoreBundle\Entity\Traits\Behavior\Blameable\BlameableTrait;
 use EightMarq\CoreBundle\Entity\Traits\Behavior\SoftDeleteable\SoftDeleteableTrait;
 use EightMarq\CoreBundle\Entity\Traits\Behavior\Timestampable\TimestampableTrait;
 use EightMarq\CoreBundle\Entity\Traits\UuidBasedEntityMethodsTrait;
-use Knp\DoctrineBehaviors\Contract\Entity\BlameableInterface;
-use Knp\DoctrineBehaviors\Model\Blameable\BlameableTrait;
 use Symfony\Bridge\Doctrine\Types\UuidType;
 use Symfony\Component\Uid\Uuid;